회원 정보 수정 시 주소 정보도 함께 업데이트 한다

+++ domain/src/main/java/sion/bookstore/domain/utils/FileUploadUtil.java
@@ -57,6 +57,10 @@ public class FileUploadUtil {
     }
 
     public void deleteExistingFile(String originalImagePath) {
+        if (Objects.isNull(originalImagePath)) {
+            return;
+        }
+
         File file = new File(originalImagePath);
         if (file.exists()) {
             file.delete();
